backend/requirements.txt
# Backend Requirements for KTU AI Assistant
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain and AI components
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0
langchain-ollama==0.1.0
langchain-chroma==0.1.0

# Vector database and embeddings
chromadb==0.4.18

# Document processing
pypdf==4.0.1

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client and utilities
requests==2.31.0
ollama==0.1.7

# Optional: Development dependencies
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
